refactor the open62541 code generation backend into a separate module
